requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.5.3
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-celery==3.2.2
flower==0.9.2
idna==2.6
kombu==3.0.37
mysqlclient==1.3.12
pika==0.11.2
pytz==2018.3
redis==2.10.6
requests==2.18.4
schedule==0.5.0
tornado==4.5.3
urllib3==1.22
vine==1.1.4
